Move feature settings into features submodule
